gzz gzzz gzzz gzzz gzzz

+++ src/cpu/ram/io.rs
@@ -49,7 +49,7 @@ pub enum Interrupt{
     TimerOverflow,
     SerialTransfer,
     Joypad,
-    AudioSample(u8,u8),
+    AudioSample(f32,f32),
 }
 
 pub struct InterruptManager{
